Extract much components to PreparePasteButton
Further simplification of Toolbar. Also removes notification from
responsibilities of PreparePasteButton / FallbackPasteModal.

To make this a fully generic "paste preparation" component, the fallback
modal text would need to be passed as a parameter.
