+++ src/SessionPage.js
@@ -1,7 +1,6 @@
 import React, { useContext, useRef, useState } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
 import { useToast } from './ToastContext';
-import { Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap'
 
 import {
   ConfigContext,
@@ -13,6 +12,7 @@ import {
 } from 'flight-webapp-components';
 
 import NoVNC from './NoVNC';
+import PreparePasteButton from './PreparePasteButton';
 import TerminateButton from './TerminateButton';
 import WrappedScreenshot from './Screenshot';
 import styles from './NoVNC.module.css';
@@ -196,11 +196,6 @@ function Toolbar({
   vnc,
 }) {
   const { addToast } = useToast();
-  const [showFallback, setShowFallback] = useState(false);
-
-  const toggleFallback = function() {
-    setShowFallback(!showFallback);
-  }
 
   const disconnectBtn = connectionState === 'connected' ? (
     <button
@@ -234,53 +229,34 @@ function Toolbar({
 
   const fullscreenBtn = <FullscreenButton onZenChange={onZenChange} />;
 
-  function handleFallbackPaste(text) {
-    try {
-      vnc.current.setClipboardText(text);
-    } catch (e) {
-      console.log('Fallback failed. Giving up.', e);  // eslint-disable-line no-console
-      const body = (
-        <div>
-          An unexpected error has occurred whilst preparing paste.  Please
-          contact your system administrator for further assistance.
-        </div>
-      );
-      addToast({body, icon: 'danger', header: 'Unexpected error' });
-    }
+  function handlePaste(text) {
+    vnc.current.setClipboardText(text);
+    const body = (
+      <div>
+        Your session's clipboard has been updated. You can now paste
+        normally within your session.
+      </div>
+    );
+    addToast({body: body, icon: 'success', header: 'Paste prepared'});
+  }
+
+  function handleFallbackError() {
+    console.log('Giving up.');  // eslint-disable-line no-console
+    const body = (
+      <div>
+        An unexpected error has occurred whilst preparing paste.  Please
+        contact your system administrator for further assistance.
+      </div>
+    );
+    addToast({body, icon: 'danger', header: 'Unexpected error' });
   }
 
   const pasteButton = connectionState === 'connected' ? (
-    <React.Fragment>
-      <button
-        className="btn btn-sm btn-light"
-        onClick={async () => {
-          try {
-            const text = await navigator.clipboard.readText();
-            if (text !== "" && vnc.current) {
-              vnc.current.setClipboardText(text);
-              const body = (
-                <div>
-                  Your session's clipboard has been updated. You can now paste
-                  normally within your session.
-                </div>
-              );
-              addToast({body: body, icon: 'success', header: 'Paste prepared'});
-            }
-          } catch (e) {
-            console.log('Paste failed. Attempting fallback.', e);  // eslint-disable-line no-console
-            toggleFallback();
-          }
-        }}
-      >
-        <i className="fa fa-paste mr-1"></i>
-        Prepare paste
-      </button>
-      <FallbackPasteModal
-        isOpen={showFallback}
-        toggle={toggleFallback}
-        onPaste={handleFallbackPaste}
-      />
-    </React.Fragment>
+    <PreparePasteButton
+      onPaste={handlePaste}
+      onFallbackError={handleFallbackError}
+      onFallbackPaste={handlePaste}
+    />
   ) : null;
 
   return (
@@ -335,42 +311,6 @@ function Screenshot({ id }) {
       className={`d-block m-auto ${styles.NoVNCWrapper}`}
       session={{ id }}
     />
-  );
-}
-
-function FallbackPasteModal({isOpen, toggle, onPaste}) {
-  const textRef = useRef();
-
-  return (
-    <Modal isOpen={isOpen} toggle={toggle}>
-      <ModalHeader toggle={toggle}>Paste Text</ModalHeader>
-      <ModalBody>
-        <p>
-          To allow your desktop session to gain access to the pasted text,
-          paste your text in the text area below and click "OK".
-        </p>
-        <p>
-          The pasted text will be added to the clipboard within your desktop
-          session and you will be able to paste normally from within your
-          session.
-        </p>
-        <textarea ref={textRef} style={{ width: "100%", height: "7em" }}></textarea>
-      </ModalBody>
-      <ModalFooter>
-        <button
-          className="btn btn-primary"
-          onClick={() => {
-            onPaste(textRef.current.value);
-            toggle();
-          }}
-        >
-          OK
-        </button>
-        <button className="btn btn-link" onClick={toggle}>
-          Cancel
-        </button>
-      </ModalFooter>
-    </Modal>
   );
 }
